integrated the api routes to use the nextGenConfig page(untested)
